Moved some functionality to results iterator
