--- src/main/java/com/awslabs/aws/greengrass/provisioner/interfaces/helpers/JsonHelper.java
@@ -1,7 +1,0 @@
-package com.awslabs.aws.greengrass.provisioner.interfaces.helpers;
-
-public interface JsonHelper {
-    String toJson(Object object);
-
-    <T> T fromJson(Class<T> clazz, byte[] json);
-}
